Change to Java
Stuck on solution in C and progress has become incredibly slow, have
changed approach to use the much more convinient Java Lingpipe package
to handle natural language approach so that all the work can be
completed before the deadline, i hope that the previous work will
suffice as evidence that i understand the process behind the package.
Currently the program searches through each item in the file and trains
the lingpipe model (which is about as far as the attempt in C).